refactor(web): extract useInstanceData hook from InstanceCard
The dashboard and inspect InstanceCard components shared an identical
data preamble (device names, counters, agent usage, memory totals).
Move it into a shared useInstanceData hook; both cards keep their own
layouts.
